Moved layer-specific tests into layers.py
